+++ packages/lumx-vue/src/index.ts
@@ -23,6 +23,7 @@ export * from './components/radio-button';
 export * from './components/skeleton';
 export * from './components/switch';
 export * from './components/text-field';
+export * from './components/uploader';
 export * from './components/table';
 export * from './components/text';
 export * from './components/thumbnail';