+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `@lumx/vue`:
+    -   Create the `Uploader` component
+
 ### Changed
 
 -   `@lumx/core`:
     -   Moved `Uploader` stories and tests from `@lumx/react`
+    -   Made `Uploader` `fileInputProps` type framework-agnostic
 
 ## [4.7.1][] - 2026-03-10